// Frontend/src/api/model/routeResponseDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface RouteResponseDTO {
  id: number;
  hallenId: number;
  wandNr: number;
  name?: string;
  farbe?: string;
  schwierigkeit?: number;
  isToprope?: boolean;
  isVorstieg?: boolean;
  schrauber?: string;
  schraubdatum?: string;
  isActive?: boolean;
  beschreibung?: string;
}